+++ src/main/java/net/minecraft/client/multiplayer/ClientLevel.java
@@ -519,7 +519,7 @@ public class ClientLevel extends Level implements BlockAndTintGetter, CacheSlot.
     }
 
     @Override
-    public boolean hasChunk(final long chunkX, final long chunkZ) {
+    public boolean hasChunk(final int chunkX, final int chunkZ) {
         return true;
     }
 
@@ -558,8 +558,7 @@ public class ClientLevel extends Level implements BlockAndTintGetter, CacheSlot.
         this.minecraft.getPlayerSocialManager().getPresenceHandler().tryUpdatePresence();
     }
 
-    // MCRe NoiseFarlands: 世界坐标 Long 化
-    public void animateTick(final long xt, final long yt, final long zt) {
+    public void animateTick(final int xt, final int yt, final int zt) {
         int r = 32;
         RandomSource animateRandom = RandomSource.createThreadLocalInstance();
         Block markerParticleTarget = this.getMarkerParticleTarget();
@@ -584,19 +583,17 @@ public class ClientLevel extends Level implements BlockAndTintGetter, CacheSlot.
     }
 
     public void doAnimateTick(
-        // MCRe NoiseFarlands: 世界坐标 Long 化
-        final long xt,
-        final long yt,
-        final long zt,
+        final int xt,
+        final int yt,
+        final int zt,
         final int r,
         final RandomSource animateRandom,
         final @Nullable Block markerParticleTarget,
         final BlockPos.MutableBlockPos pos
     ) {
-        // MCRe NoiseFarlands: 世界坐标 Long 化（随机偏移后仍为世界坐标）
-        long x = xt + this.random.nextInt(r) - this.random.nextInt(r);
-        long y = yt + this.random.nextInt(r) - this.random.nextInt(r);
-        long z = zt + this.random.nextInt(r) - this.random.nextInt(r);
+        int x = xt + this.random.nextInt(r) - this.random.nextInt(r);
+        int y = yt + this.random.nextInt(r) - this.random.nextInt(r);
+        int z = zt + this.random.nextInt(r) - this.random.nextInt(r);
         pos.set(x, y, z);
         BlockState state = this.getBlockState(pos);
         state.getBlock().animateTick(state, this, pos, animateRandom);
@@ -843,13 +840,12 @@ public class ClientLevel extends Level implements BlockAndTintGetter, CacheSlot.
         this.levelExtractor.setBlockDirty(pos, oldState, newState);
     }
 
-    public void setSectionDirtyWithNeighbors(final long chunkX, final long chunkY, final long chunkZ) {
+    public void setSectionDirtyWithNeighbors(final int chunkX, final int chunkY, final int chunkZ) {
         this.levelExtractor.setSectionDirtyWithNeighbors(chunkX, chunkY, chunkZ);
     }
 
-    // MCRe NoiseFarlands: section 世界坐标 Long 化
     public void setSectionRangeDirty(
-        final long minSectionX, final long minSectionY, final long minSectionZ, final long maxSectionX, final long maxSectionY, final long maxSectionZ
+        final int minSectionX, final int minSectionY, final int minSectionZ, final int maxSectionX, final int maxSectionY, final int maxSectionZ
     ) {
         this.levelExtractor.setSectionRangeDirty(minSectionX, minSectionY, minSectionZ, maxSectionX, maxSectionY, maxSectionZ);
     }
@@ -995,7 +991,7 @@ public class ClientLevel extends Level implements BlockAndTintGetter, CacheSlot.
     }
 
     @Override
-    public Holder<Biome> getUncachedNoiseBiome(final long quartX, final long quartY, final long quartZ) {
+    public Holder<Biome> getUncachedNoiseBiome(final int quartX, final int quartY, final int quartZ) {
         return this.registryAccess().lookupOrThrow(Registries.BIOME).getOrThrow(Biomes.PLAINS);
     }
 
@@ -1125,9 +1121,9 @@ public class ClientLevel extends Level implements BlockAndTintGetter, CacheSlot.
     public void addBreakingBlockEffect(final BlockPos pos, final Direction direction) {
         BlockState blockState = this.getBlockState(pos);
         if (blockState.getRenderShape() != RenderShape.INVISIBLE && blockState.shouldSpawnTerrainParticles()) {
-            long x = pos.getX();
-            long y = pos.getY();
-            long z = pos.getZ();
+            int x = pos.getX();
+            int y = pos.getY();
+            int z = pos.getZ();
             float r = 0.1F;
             AABB shape = blockState.getShape(this, pos).bounds();
             double xp = x + this.random.nextDouble() * (shape.maxX - shape.minX - 0.2F) + 0.1F + shape.minX;
